Move createOrAddListConfig
