+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/AbstractConfigurationTest.java
@@ -14,12 +14,15 @@ import io.vertx.core.json.JsonObject;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.stream.Stream;
 
 import static io.strimzi.test.TestUtils.LINE_SEPARATOR;
 import static java.util.Arrays.asList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.fail;
 
 @ParallelSuite
@@ -246,6 +249,91 @@ public class AbstractConfigurationTest {
         assertThat(configuration.asOrderedProperties().asMap().get("key.converter"), is("my.package.Converter"));
         assertThat(configuration.asOrderedProperties().asMap().get("ssl.keystore.location"), is(nullValue()));
         assertThat(configuration.asOrderedProperties().asMap().get("ssl.endpoint.identification.algorithm"), is(""));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigDoesNotExists() {
+        Map<String, Object> userConfiguration = new HashMap<>();
+        KafkaConnectConfiguration config1 = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        AbstractConfiguration.createOrAddListConfig(config1, "test-key", "test-value");
+        assertThat(config1.getConfigOption("test-key"), is("test-value"));
+
+        KafkaConnectConfiguration config2 = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        AbstractConfiguration.createOrAddListConfig(config2, "test-key", "test-value-1,test-value-2");
+        assertThat(config2.getConfigOption("test-key"), is("test-value-1,test-value-2"));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigExists() {
+        Map<String, Object> userConfiguration = new HashMap<>();
+        KafkaConnectConfiguration config1 = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        config1.setConfigOption("test-key", "test-value-1");
+
+        AbstractConfiguration.createOrAddListConfig(config1, "test-key", "test-value-2");
+        assertThat(config1.getConfigOption("test-key"), is("test-value-1,test-value-2"));
+
+        KafkaConnectConfiguration config2 = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        config2.setConfigOption("test-key", "test-value-1,test-value-2");
+        AbstractConfiguration.createOrAddListConfig(config2, "test-key", "test-value-3");
+        assertThat(config2.getConfigOption("test-key"), is("test-value-1,test-value-2,test-value-3"));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigExistsAndContainsValue() {
+        Map<String, Object> userConfiguration = new HashMap<>();
+        KafkaConnectConfiguration config1 = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        config1.setConfigOption("test-key", "test-value-1");
+        AbstractConfiguration.createOrAddListConfig(config1, "test-key", "test-value-1");
+        assertThat(config1.getConfigOption("test-key"), is("test-value-1"));
+
+        KafkaConnectConfiguration config2 = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        config2.setConfigOption("test-key", "test-value-1,test-value-2");
+        AbstractConfiguration.createOrAddListConfig(config2, "test-key", "test-value-1");
+        assertThat(config2.getConfigOption("test-key"), is("test-value-1,test-value-2"));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigWithDuplicates() {
+        Map<String, Object> userConfiguration = new HashMap<>();
+        KafkaConnectConfiguration config = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        config.setConfigOption("test-key", "test-value-1,test-value-1,test-value-2");
+        AbstractConfiguration.createOrAddListConfig(config, "test-key", "test-value-3,test-value-3");
+        assertThat(config.getConfigOption("test-key"), is("test-value-1,test-value-2,test-value-3"));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigOrdering() {
+        Map<String, Object> userConfiguration = new HashMap<>();
+        KafkaConnectConfiguration config = new KafkaConnectConfiguration(Reconciliation.DUMMY_RECONCILIATION, userConfiguration.entrySet());
+        config.setConfigOption("test-key", "test-value-2,test-value-1");
+        AbstractConfiguration.createOrAddListConfig(config, "test-key", "test-value-3,");
+        assertThat(config.getConfigOption("test-key"), is("test-value-2,test-value-1,test-value-3"));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigWithNullConfig() {
+        KafkaConfiguration config = new KafkaConfiguration(Reconciliation.DUMMY_RECONCILIATION, Set.of());
+        config.setConfigOption("test-key", "test-value-1,test-value-2");
+        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> AbstractConfiguration.createOrAddListConfig(null, "test-key", "test-value-3"));
+        assertThat(exception.getMessage(), is("Configuration is required"));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigWithNullKey() {
+        KafkaConfiguration config = new KafkaConfiguration(Reconciliation.DUMMY_RECONCILIATION, Set.of());
+        config.setConfigOption("test-key", "test-value-1,test-value-2");
+        Stream.of(null, "", " ")
+                .map(key -> assertThrows(IllegalArgumentException.class, () -> AbstractConfiguration.createOrAddListConfig(config, key, "test-value-3")))
+                .forEach(e -> assertThat(e.getMessage(), is("Configuration key is required")));
+    }
+
+    @ParallelTest
+    public void testCreateOrAddListConfigWithNullValue() {
+        KafkaConfiguration config = new KafkaConfiguration(Reconciliation.DUMMY_RECONCILIATION, Set.of());
+        config.setConfigOption("test-key", "test-value-1,test-value-2");
+        Stream.of(null, "", " ")
+                .map(value -> assertThrows(IllegalArgumentException.class, () -> AbstractConfiguration.createOrAddListConfig(config, "test-key", value)))
+                .forEach(e -> assertThat(e.getMessage(), is("Configuration values are required")));
     }
 
     @ParallelTest
